+++ src/main/java/com/RoutineGongJakSo/BE/security/validator/Validator.java
@@ -1,6 +1,7 @@
 package com.RoutineGongJakSo.BE.security.validator;
 
 import com.RoutineGongJakSo.BE.model.User;
+import com.RoutineGongJakSo.BE.model.WeekTeam;
 import com.RoutineGongJakSo.BE.repository.UserRepository;
 import com.RoutineGongJakSo.BE.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +12,6 @@ import java.util.Optional;
 @Component
 @RequiredArgsConstructor
 public class Validator {
-    public final UserRepository userRepository;
 
     //유저 아이디 중복확인
     public static void checkUser(Optional<User> found) {
@@ -32,5 +32,13 @@ public class Validator {
         if (userDetails == null) {
             throw new NullPointerException("로그인 된 사용자가 아닙니다.");
         }
+    }
+
+    //이미 만들어진 팀이 있는지 확인
+    public void teamCheck(Optional<WeekTeam> teamCheck){
+        if (teamCheck.isPresent()){
+            throw new NullPointerException("이미 만들어진 팀이 존재합니다.");
+        }
+
     }
 }
